refactor: change grid component architecture
